Created utility.js for the maybe frequently used functions. Fixed a few small style issues.
