+++ src/client/editor/log.js
@@ -1,3 +1,5 @@
+import { currentTimeString } from "../../utility";
+
 let element;
 
 export function setEditorLogElement(el) {
@@ -5,22 +7,9 @@ export function setEditorLogElement(el) {
 }
 
 export function editorLog(message, type = "info") {
-	let date = new Date();
-	let hrs = date.getHours();
-	let min = date.getMinutes();
-	let sec = date.getSeconds();
-	if (hrs < 10) {
-		hrs = `0${hrs}`;
-	}
-	if (min < 10) {
-		min = `0${min}`;
-	}
-	if (sec < 10) {
-		sec = `0${sec}`;
-	}
 	element.insertAdjacentHTML(
 		"beforeend",
-		`<div class='${type}'>[${hrs}:${min}:${sec}] ${message}</div>`
+		`<div class='${type}'>[${currentTimeString()}] ${message}</div>`
 	);
 	element.scrollTop = element.scrollHeight;
 }
