+++ src/log.js
@@ -1,27 +1,19 @@
+const utility = require("./utility.js");
+
 module.exports = {
 	info: function(message, ...args) {
-		console.log(`${currentHourString()}${message}`, ...args);
+		console.log(`[${utility.currentTimeString()}] ${message}`, ...args);
 	},
 
 	warn: function(message, ...args) {
-		console.log(`${currentHourString().yellow}${message}`, ...args);
+		console.log(`[${utility.currentTimeString().yellow}] ${message}`, ...args);
 	},
 
 	error: function(message, ...args) {
-		console.log(`${currentHourString().red}${message}`, ...args);
+		console.log(`[${utility.currentTimeString().red}] ${message}`, ...args);
 	},
 
 	throw: function(message) {
-		throw `${currentHourString().red}${"[FATAL]".red} ${message}`;
+		throw `[${utility.currentTimeString().red}] ${"[FATAL]".red} ${message}`;
 	}
 };
-
-function pad(num, size) {
-	let s = `000000000${num}`;
-	return s.substr(s.length - size);
-}
-
-function currentHourString() {
-	let date = new Date();
-	return `[${pad(date.getHours(), 2)}:${pad(date.getMinutes(), 2)}] `;
-}
